+++ tests/test_stac_api_io.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from urllib.parse import parse_qs, urlsplit
 
 import pytest
+import pystac
 from pytest import MonkeyPatch
 from requests_mock.mocker import Mocker
 
@@ -274,3 +275,15 @@ class TestSTAC_IOOverride:
         stac_api_io = StacApiIO()
         with pytest.raises(APIError):
             stac_api_io.request("https://earth-search.aws.element84.com/v1/")
+
+
+@pytest.mark.vcr
+def test_stac_io_in_pystac() -> None:
+    # https://github.com/stac-utils/pystac-client/issues/706
+    stac_io = StacApiIO(timeout=42)
+    collection = pystac.read_file(
+        href="tests/data/planetary-computer-collection.json",
+        stac_io=stac_io,
+    )
+    stac_io = collection.get_root()._stac_io
+    assert stac_io.timeout == 42